Migrate check_link_target_js to BookTest
